@@ -1,47 +1,50 @@
-import React, { useState } from 'react';
-import './App.css';
-import { useDispatch } from 'react-redux';
-import { USER_TYPE } from './store/UserReducer';
+import React, { useState } from "react";
+import { useDispatch } from "react-redux";
+import "./App.css";
+import { USER_TYPE } from "./store/UserReducer";
 import UserDisplay from "./UserDisplay";
 import PostDisplay from "./PostDisplay";
-import { POST_TYPE } from './store/PostReducer';
+import { POST_TYPE } from "./store/PostReducer";
 
 function App() {
   const [userid, setUserid] = useState(0);
   const dispatch = useDispatch();
   const [postid, setPostId] = useState(0);
-
   const onChangeUserId = async (e: React.ChangeEvent<HTMLInputElement>) => {
     const useridFromInput = e.target.value ? Number(e.target.value) : 0;
     console.log("userid", useridFromInput);
     setUserid(useridFromInput);
 
-    const usersResponse = await fetch('https://jsonplaceholder.typicode.com/users');
-    if(usersResponse.ok) {
+    const usersResponse = await fetch(
+      "https://jsonplaceholder.typicode.com/users"
+    );
+    if (usersResponse.ok) {
       const users = await usersResponse.json();
-      
+      console.log("users", users);
       const usr = users.find((userItem: any) => {
         return userItem && userItem.id === useridFromInput;
       });
-      
+      console.log("usr", usr);
       dispatch({
         type: USER_TYPE,
         payload: {
           id: usr.id,
           username: usr.username,
           email: usr.email,
-          city: usr.address.city
-        }
+          city: usr.address.city,
+        },
       });
     }
-  }
+  };
 
   const onChangePostId = async (e: React.ChangeEvent<HTMLInputElement>) => {
     const postIdFromInput = e.target.value ? Number(e.target.value) : 0;
     setPostId(postIdFromInput);
 
-    const postResponse = await fetch("https://jsonplaceholder.typicode.com/posts/" + postIdFromInput);
-    if(postResponse.ok) {
+    const postResponse = await fetch(
+      "https://jsonplaceholder.typicode.com/posts/" + postIdFromInput
+    );
+    if (postResponse.ok) {
       const post = await postResponse.json();
       console.log("post", post);
       dispatch({
@@ -49,23 +52,23 @@ function App() {
         payload: {
           id: post.id,
           title: post.title,
-          body: post.body
-        }
-      })
+          body: post.body,
+        },
+      });
     }
-  }
+  };
 
   return (
     <React.Fragment>
-      <div style={{width: "300px"}}>
+      <div style={{ width: "300px" }}>
         <div className="App">
           <label>user id</label>
           <input value={userid} onChange={onChangeUserId} />
         </div>
         <UserDisplay />
       </div>
-      <br/>
-      <div style={{width: "300px"}}>
+      <br />
+      <div style={{ width: "300px" }}>
         <div className="App">
           <label>post id</label>
           <input value={postid} onChange={onChangePostId} />
