@@ -1,12 +1,9 @@
-import React, { useContext } from 'react';
-import { TestContext } from './ContextTester';
+import React, { useContext } from "react";
+import { TestContext } from "./ContextTester";
 
 const UserAgeComp = () => {
-    const { userage } = useContext(TestContext);
-    
-    return <div>
-        {userage}
-    </div>
+  const { userage } = useContext(TestContext);
+  return <div>{userage}</div>;
 };
 
 export default UserAgeComp;
